+++ src/tests/mod.rs
@@ -1,1 +1,2 @@
 ﻿pub mod server_tests;
+mod load_tests;